The program should identify those characters which differ between the two given strings in a visually striking way.
Output the two input strings on two lines, and then identify the differences on the line below using periods (for identical characters) and asterisks (for differing characters).

For example:

ATCCGCTTAGAGGGATT
GTCCGTTTAGAAGGTTT
*....*.....*..*..

The first line of input contains an integer 1≤n≤500, indicating the number of test cases that follow. Each test case is a pair of lines of the same length, 1 to 50 characters. Each string contains only letters (a-z, A-Z) or digits (0-9).